+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"25ffdf2c-9ead-4bf8-9ff6-5739da432b80":{"parent":null,"children":[]}},"files":{"D:\\courses\\Andela\\Broadcaster\\server\\models\\users.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\app.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\routes\\authRouter.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\controlers\\authController.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\helper\\tokenGen.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\helper\\passwordHash.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\middleware\\validation.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\routes\\recordRouter.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\middleware\\filefolder.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\controlers\\recordController.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\models\\record.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"],"D:\\courses\\Andela\\Broadcaster\\server\\helper\\tokenVerify.js":["25ffdf2c-9ead-4bf8-9ff6-5739da432b80"]},"externalIds":{}}
+{"processes":{"30206d8a-ebc0-4ede-961a-6378d654f9c9":{"parent":null,"children":[]}},"files":{"D:\\courses\\Andela\\Broadcaster\\server\\models\\users.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\app.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\routes\\authRouter.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\controlers\\authController.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\helper\\tokenGen.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\helper\\passwordHash.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\middleware\\validation.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\routes\\recordRouter.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\middleware\\filefolder.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\controlers\\recordController.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\models\\record.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"],"D:\\courses\\Andela\\Broadcaster\\server\\helper\\tokenVerify.js":["30206d8a-ebc0-4ede-961a-6378d654f9c9"]},"externalIds":{}}
