+++ package.json
@@ -7,7 +7,7 @@
     "start": "npm run build && node build/app.js",
     "build": "babel ./server -d build",
     "dev": "nodemon --exec babel-node ./server/app.js",
-    "test": "nyc --reporter=html --reporter=text mocha --require @babel/register --timeout 10000 --exit server/test/user.test.js ",
+    "test": "nyc --reporter=html --reporter=text mocha --require @babel/register --timeout 10000 --exit server/test/*.test.js ",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
